Added icons
Implemented icons and icon logic. Added logic for showing the icon on a slot + the quantity of the item. When an icon is hovered, it glows white for selection. If there is no item associated with a slot, there is no icon or quantity count. UP NEXT: finish inventory UI to show all slots with items