Finished ItemInfoUI
The UI for displaying item info when hovered in inventory screen is complete. When an item is hovered, it shows the stats, name, and description of the item. When an empty slot is hovered, no item info is shown. Item info is also shown along with the pick up UI button when looking at an interactable item.