Added the Gemstone System
Gems will be extremely powerful items that significantly boost players stats. Implemented all logic for another inventory tab for gems, as well as the logic for equipping gems. Also implemented logic for blue, green and red metallic materials for the gemstone colors. Changed Z order for the character model and slot buttons so that there is no issues when clicking on item slots.