Controller inputs for UI and updates
Made an attempt for an implementation of using a controller to move the mouse cursor within UI components - Unsuccessful. Going to give it another go at a later time. Updated blueprints for the player to use ESC key to exit inventory. Changed opening inventory so that the game pauses to avoid performance issues. Added main menu picture